requirements.txt
Django==1.5.12
boto3==1.4.7
requests==2.18.4
django_graphos==0.3.41
futures==3.2.0
Pillow==5.0.0
django-graphos==0.3.41